+++ vitest.config.ts
@@ -8,6 +8,6 @@ export default defineConfig({
     // TODO: Parallelize tests and mark integration tests as long-running
     testTimeout: 120000,
     hookTimeout: 240000,
-    threads: false // this is needed to do process.chdir() in tests. It causes the tests to not be isolated from one another.
+    pool: 'forks' // this is needed to do process.chdir() in tests.
   }
 });