+++ packages/importer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @xata.io/importer
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [[`e97d1999`](https://github.com/xataio/client-ts/commit/e97d1999f3c25f149213ceca81958e1674624e05)]:
+  - @xata.io/client@0.28.0
+
 ## 1.1.2
 
 ### Patch Changes